example-structure/requirements.txt
# requirements.txt
# Simple dependency list for the instrument control software
# Install with: pip install -r requirements.txt

numpy>=1.20.0
pyvisa>=1.13.0
pyvisa-sim>=0.5.0
pyserial>=3.5
pydantic>=2.0.0
pyyaml>=6.0
qtpy>=2.0.0
pyqt5>=5.15.0

# Optional development dependencies
# pytest>=7.0.0
# pytest-qt>=4.2.0
# black>=23.0.0

